requirements/dev.txt
-r common.txt
coverage
django-extensions
werkzeug